remade the kiss libarary as a C++ class, it now builds
