+++ Builds/VisualStudio2019/MidiVisualizer_ConsoleApp.vcxproj
@@ -865,11 +865,8 @@
     <ClCompile Include="..\..\JuceLibraryCode\include_juce_audio_devices.cpp" />
     <ClCompile Include="..\..\JuceLibraryCode\include_juce_core.cpp" />
     <ClCompile Include="..\..\JuceLibraryCode\include_juce_events.cpp" />
-    <ClCompile Include="..\..\Source\kiss_draw.c" />
-    <ClCompile Include="..\..\Source\kiss_general.c" />
-    <ClCompile Include="..\..\Source\kiss_posix.c" />
-    <ClCompile Include="..\..\Source\kiss_widgets.c" />
     <ClCompile Include="..\..\Source\Main.cpp" />
+    <ClCompile Include="..\..\Source\myKissGui.cpp" />
     <ClCompile Include="..\..\Source\RenderWindow.cpp" />
     <ClCompile Include="..\..\Source\Translator.cpp" />
     <ClCompile Include="..\..\Source\UserGUI.cpp" />
@@ -1181,8 +1178,8 @@
     <ClInclude Include="..\..\..\..\JUCE\modules\juce_events\juce_events.h" />
     <ClInclude Include="..\..\JuceLibraryCode\JuceHeader.h" />
     <ClInclude Include="..\..\Source\Binding.hpp" />
-    <ClInclude Include="..\..\Source\kiss_sdl.h" />
     <ClInclude Include="..\..\Source\midiModule.h" />
+    <ClInclude Include="..\..\Source\myKissGui.hpp" />
     <ClInclude Include="..\..\Source\RenderWindow.hpp" />
     <ClInclude Include="..\..\Source\Sprite.hpp" />
     <ClInclude Include="..\..\Source\UserGUI.hpp" />
