+++ Builds/VisualStudio2019/x64/Debug/ConsoleApp/MidiVisualizer.exe.recipe
@@ -1,7 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project>
-  <ProjectOutputs>C:\Users\qwalt\source\repos\CS595\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualizer.exe</ProjectOutputs>
-  <ContentFiles></ContentFiles>
-  <SatelliteDlls></SatelliteDlls>
-  <NonRecipeFileRefs></NonRecipeFileRefs>
+  <ProjectOutputs>
+    <ProjectOutput>
+      <FullPath>C:\Files\CS595\Repo\CS595Group12\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualizer.exe</FullPath>
+    </ProjectOutput>
+  </ProjectOutputs>
+  <ContentFiles />
+  <SatelliteDlls />
+  <NonRecipeFileRefs />
 </Project>
