+++ Builds/VisualStudio2019/x64/Debug/ConsoleApp/MidiVisualizer_ConsoleApp.vcxproj.FileListAbsolute.txt
@@ -2,3 +2,4 @@ C:\Files\CS595Project\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualize
 C:\Files\CS595Project\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualizer.pdb
 C:\Users\qwalt\source\repos\CS595\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualizer.exe
 C:\Users\qwalt\source\repos\CS595\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualizer.pdb
+C:\Files\CS595\Repo\CS595Group12\Builds\VisualStudio2019\x64\Debug\ConsoleApp\MidiVisualizer.exe
