+++ Source/UserGUI.cpp
@@ -6,10 +6,12 @@
 #include <vector>
 
 #include "UserGUI.hpp"
-#include "kiss_sdl.h"
+//#include "kiss_sdl.h"
+#include "myKissGui.hpp"
 
 using std::cout; using std::cin;
 using std::endl; using std::string;
+
 
 UserGUI::UserGUI(const char* p_title) :mode(), window(NULL), renderer(NULL)
 {
@@ -31,6 +33,7 @@ UserGUI::UserGUI(const char* p_title) :mode(), window(NULL), renderer(NULL)
 	}
 
 	cout << "GUI strated" << endl;
+	myKissGUI* kissGUI = new myKissGUI();
 
 	//initializeScene();
 }
@@ -56,6 +59,7 @@ void UserGUI::enterViewMode()
 	 and any time the directory changes. */
 int UserGUI::initializeScene()
 {
+	
 	kiss_array objects, a1, a2;
 	kiss_window window1, window2;
 	kiss_label label1 = { 0 }, label2 = { 0 }, label_sel = { 0 },
@@ -74,55 +78,57 @@ int UserGUI::initializeScene()
 	textbox_height = 250;
 	window2_width = 400;
 	window2_height = 168;
-	renderer = kiss_init("kiss_sdl example 1", &objects, 640, 480);
+	renderer = kissGUI->kiss_init("kiss_sdl example 1", &objects, 640, 480);
 	if (!renderer) return 1;
-	kiss_array_new(&a1);
-	kiss_array_append(&objects, ARRAY_TYPE, &a1);
-	kiss_array_new(&a2);
-	kiss_array_append(&objects, ARRAY_TYPE, &a2);
+	kissGUI->kiss_array_new(&a1);
+	kissGUI->kiss_array_append(&objects, ARRAY_TYPE, &a1);
+	kissGUI->kiss_array_new(&a2);
+	kissGUI->kiss_array_append(&objects, ARRAY_TYPE, &a2);
 
 	/* Arrange the widgets nicely relative to each other */
-	kiss_window_new(&window1, NULL, 1, 0, 0, 500,
+	kissGUI->kiss_window_new(&window1, NULL, 1, 0, 0, 500,
 		500);
-	/*kiss_textbox_new(&textbox1, &window1, 1, &a1, kiss_screen_width / 2 -
-		(2 * textbox_width + 2 * kiss_up.w - kiss_edge) / 2,
-		3 * kiss_normal.h, textbox_width, textbox_height);
-	kiss_vscrollbar_new(&vscrollbar1, &window1, textbox1.rect.x +
+	
+	kissGUI->kiss_textbox_new(&textbox1, &window1, 1, &a1, kissGUI->kiss_screen_width / 2 -
+		(2 * textbox_width + 2 * kissGUI->kiss_up.w - kissGUI->kiss_edge) / 2,
+		3 * kissGUI->kiss_normal.h, textbox_width, textbox_height);
+	
+	kissGUI->kiss_vscrollbar_new(&vscrollbar1, &window1, textbox1.rect.x +
 		textbox_width, textbox1.rect.y, textbox_height);
-	kiss_textbox_new(&textbox2, &window1, 1, &a2,
-		vscrollbar1.uprect.x + kiss_up.w, textbox1.rect.y,
+	kissGUI->kiss_textbox_new(&textbox2, &window1, 1, &a2,
+		vscrollbar1.uprect.x + kissGUI->kiss_up.w, textbox1.rect.y,
 		textbox_width, textbox_height);
-	kiss_vscrollbar_new(&vscrollbar2, &window1, textbox2.rect.x +
+	kissGUI->kiss_vscrollbar_new(&vscrollbar2, &window1, textbox2.rect.x +
 		textbox_width, vscrollbar1.uprect.y, textbox_height);
-	kiss_label_new(&label1, &window1, "Folders", textbox1.rect.x +
-		kiss_edge, textbox1.rect.y - kiss_textfont.lineheight);
-	kiss_label_new(&label2, &window1, "Files", textbox2.rect.x +
-		kiss_edge, textbox1.rect.y - kiss_textfont.lineheight);
-	kiss_label_new(&label_sel, &window1, "", textbox1.rect.x +
-		kiss_edge, textbox1.rect.y + textbox_height +
-		kiss_normal.h);
-	kiss_entry_new(&entry, &window1, 1, "kiss", textbox1.rect.x,
-		label_sel.rect.y + kiss_textfont.lineheight,
-		2 * textbox_width + 2 * kiss_up.w + kiss_edge);
-	kiss_button_new(&button_cancel, &window1, "Cancel",
-		entry.rect.x + entry.rect.w - kiss_edge - kiss_normal.w,
-		entry.rect.y + entry.rect.h + kiss_normal.h);
-	kiss_button_new(&button_ok1, &window1, "OK", button_cancel.rect.x -
-		2 * kiss_normal.w, button_cancel.rect.y);
-	kiss_window_new(&window2, NULL, 1, kiss_screen_width / 2 -
-		window2_width / 2, kiss_screen_height / 2 -
+	kissGUI->kiss_label_new(&label1, &window1, "Folders", textbox1.rect.x +
+		kissGUI->kiss_edge, textbox1.rect.y - kissGUI->kiss_textfont.lineheight);
+	kissGUI->kiss_label_new(&label2, &window1, "Files", textbox2.rect.x +
+		kissGUI->kiss_edge, textbox1.rect.y - kissGUI->kiss_textfont.lineheight);
+	kissGUI->kiss_label_new(&label_sel, &window1, "", textbox1.rect.x +
+		kissGUI->kiss_edge, textbox1.rect.y + textbox_height +
+		kissGUI->kiss_normal.h);
+	kissGUI->kiss_entry_new(&entry, &window1, 1, "kiss", textbox1.rect.x,
+		label_sel.rect.y + kissGUI->kiss_textfont.lineheight,
+		2 * textbox_width + 2 * kissGUI->kiss_up.w + kissGUI->kiss_edge);
+	kissGUI->kiss_button_new(&button_cancel, &window1, "Cancel",
+		entry.rect.x + entry.rect.w - kissGUI->kiss_edge - kissGUI->kiss_normal.w,
+		entry.rect.y + entry.rect.h + kissGUI->kiss_normal.h);
+	kissGUI->kiss_button_new(&button_ok1, &window1, "OK", button_cancel.rect.x -
+		2 * kissGUI->kiss_normal.w, button_cancel.rect.y);
+	kissGUI->kiss_window_new(&window2, NULL, 1, kissGUI->kiss_screen_width / 2 -
+		window2_width / 2, kissGUI->kiss_screen_height / 2 -
 		window2_height / 2, window2_width, window2_height);
-	kiss_label_new(&label_res, &window2, "", window2.rect.x +
-		kiss_up.w, window2.rect.y + kiss_vslider.h);
-	label_res.textcolor = kiss_blue;
-	kiss_progressbar_new(&progressbar, &window2, window2.rect.x +
-		kiss_up.w - kiss_edge, window2.rect.y + window2.rect.h / 2 -
-		kiss_bar.h / 2 - kiss_border,
-		window2.rect.w - 2 * kiss_up.w + 2 * kiss_edge);
-	kiss_button_new(&button_ok2, &window2, "OK", window2.rect.x +
-		window2.rect.w / 2 - kiss_normal.w / 2,
+	kissGUI->kiss_label_new(&label_res, &window2, "", window2.rect.x +
+		kissGUI->kiss_up.w, window2.rect.y + kissGUI->kiss_vslider.h);
+	label_res.textcolor = kissGUI->kiss_blue;
+	kissGUI->kiss_progressbar_new(&progressbar, &window2, window2.rect.x +
+		kissGUI->kiss_up.w - kissGUI->kiss_edge, window2.rect.y + window2.rect.h / 2 -
+		kissGUI->kiss_bar.h / 2 - kissGUI->kiss_border,
+		window2.rect.w - 2 * kissGUI->kiss_up.w + 2 * kissGUI->kiss_edge);
+	kissGUI->kiss_button_new(&button_ok2, &window2, "OK", window2.rect.x +
+		window2.rect.w / 2 - kissGUI->kiss_normal.w / 2,
 		progressbar.rect.y + progressbar.rect.h +
-		2 * kiss_vslider.h);*/
+		2 * kissGUI->kiss_vslider.h);
 }
 
 /* This method will render the sprites that are held in the sprite vector to the screen.
