+++ Source/UserGUI.hpp
@@ -5,6 +5,7 @@
 #include <filesystem>
 #include <string>
 #include <vector>
+#include "myKissGui.hpp"
 
 class UserGUI
 {
@@ -17,6 +18,7 @@ public:
 	std::filesystem::path selectedDirectory;
 	SDL_DisplayMode mode;
 	SDL_Window* window;
+	myKissGUI* kissGUI;
 private:
 	SDL_Renderer* renderer;
 	SDL_Event e;
